deletion confirmation warnings added

+++ src/auth/DeleteUser.js
@@ -5,29 +5,35 @@ import { Form, FormGroup, ModalHeader, Button, Label, Input, Modal, ModalBody } 
 
 
 const Delete = (props) => {
-    console.log('User Delete Function Called');
+    const proceed = window.confirm("Are you certain you wish to delete your own user account? Your characters will remain, but you will no longer be able to access them. This is a permanent action that cannot be undone.");
+    if (proceed) {
 
-    const fetch_url = `http://localhost:3000/user/delete/loggedInUser`;
+        console.log('User Delete Function Called');
 
-    fetch(fetch_url, {
-        method: 'DELETE',
-        headers: new Headers({
-            "Content-Type": "application/json",
-            "Authorization": props.sessionToken
+        const fetch_url = `http://localhost:3000/user/delete/loggedInUser`;
+
+        fetch(fetch_url, {
+            method: 'DELETE',
+            headers: new Headers({
+                "Content-Type": "application/json",
+                "Authorization": props.sessionToken
+            })
         })
-    })
 
-    return(
-        <>
-        </>
-        // <Modal isOpen={true}>
-        //     <ModalHeader>Are you sure you want to delete your user?</ModalHeader>
-        //         <ModalBody>
-        //             <Button className="deleteBtn" type="button" onClick={DeleteUser}>Yes, Delete User</Button>
-        //             <Button className="cancelDel" type="button" onClick={props.deleteOff}>No, Go Back</Button>
-        //         </ModalBody>
-        // </Modal>
-    )
+        return(
+            <>
+            </>
+            // <Modal isOpen={true}>
+            //     <ModalHeader>Are you sure you want to delete your user?</ModalHeader>
+            //         <ModalBody>
+            //             <Button className="deleteBtn" type="button" onClick={DeleteUser}>Yes, Delete User</Button>
+            //             <Button className="cancelDel" type="button" onClick={props.deleteOff}>No, Go Back</Button>
+            //         </ModalBody>
+            // </Modal>
+        )
+    } else {
+        console.log("Delete cancelled.");
+    }
 }
 //}
